Clarify for Coverty Scan, we only use relative time
store_truncates_time_t: A time_t value is stored in an integer with too
   few bits to accommodate it. The expression ptr->time is cast to int.

Fixes CID 408300 and CID 408301

+++ src/timer.c
@@ -127,7 +127,7 @@ int timer_next_delay(void)
 	return 0;
     }
 
-    return ptr->time;
+    return (int)ptr->time;
 }
 
 /*
@@ -209,7 +209,7 @@ int timer_get(int timer_id)
     TAILQ_FOREACH(ptr, &tl, link) {
 	left += ptr->time;
 	if (ptr->id == timer_id)
-	    return left;
+	    return (int)left;
     }
 
     return -1;
